Refactor: Migrate sample app from Room to SQLDelight
This commit replaces the Room database implementation with SQLDelight in the sample project and introduces support for `Long`-keyed paging sources in the `netflow-paging` library.

The migration facilitates better multiplatform support and aligns with SQLDelight's paging extensions.

**Library Changes:**
- Introduced `LongKeyedPagingSourceWrapper` to bridge `PagingSource<Long, T>` (standard for SQLDelight) to `PagingSource<Int, T>`.
- Added `localSourceLong` to `PagingBuilder` to simplify using SQLDelight paging sources within `responsePaginated`.

**Sample App Changes:**
- Removed Room dependencies, DAOs, entities, and database configuration.
- Added SQLDelight schema and queries in `Todo.sq`.
- Updated `TodoRepositoryImpl` to use SQLDelight queries and transactions.
- Updated Koin `databaseModule` for both Android and iOS to provide `SqlDriver` and the generated `AppDatabase`.
- Adjusted `TodoMapper` to handle type conversions between SQLDelight's `Long` types and the domain model's `Int` types.
- Updated documentation in `README.md` to reflect SQLDelight usage.

+++ sample/src/commonMain/kotlin/com/kmpbits/sample/android/data/mapper/TodoMapper.kt
@@ -1,6 +1,6 @@
 package com.kmpbits.sample.android.data.mapper
 
-import com.kmpbits.sample.android.data.database.entity.TodoEntity
+import com.kmpbits.sample.android.data.database.TodoEntity
 import com.kmpbits.sample.android.data.dto.TodoDto
 import com.kmpbits.sample.android.domain.model.Todo
 
@@ -12,27 +12,27 @@ fun TodoDto.toModel() = Todo(
 )
 
 fun TodoEntity.toModel() = Todo(
-    id = id,
-    userId = userId,
+    id = id.toInt(),
+    userId = userId.toInt(),
     title = title,
     completed = completed
 )
 
 fun TodoEntity.toDto() = TodoDto(
-    id = id,
-    userId = userId,
+    id = id.toInt(),
+    userId = userId.toInt(),
     title = title,
     completed = completed
 ).also {
-    it.page = page
+    it.page = page.toInt()
     it.lastUpdatedTimestamp = lastUpdatedTimestamp
 }
 
 fun TodoDto.toEntity() = TodoEntity(
-    id = id,
-    userId = userId,
+    id = id.toLong(),
+    userId = userId.toLong(),
     title = title,
     completed = completed,
-    page = page,
+    page = page.toLong(),
     lastUpdatedTimestamp = lastUpdatedTimestamp
 )
